Moved the session notes model, schema, and mutation to scheduler. Deleted the original makemigration files. Finally OFFICIALLY done with the backend
